ref: transfer_virtual_balance reusable function, remove cross_chain_addresses from execute_transfer_voucher_internal
